Add files via upload
Debugging this code was changing the "&&" operator in the 4th line inside the for loop to "&". The problem was using && over & is one being used specifically for vectors. As seen in the code, this function works with creating true values in a vector. therefore, "&&" would not work, a "&" operator is vectorized and the correct option